split server code into library and HTTP server implementation
