+++ tools/CMakeLists.txt
@@ -21,6 +21,7 @@ else()
     add_subdirectory(main)
     add_subdirectory(perplexity)
     add_subdirectory(quantize)
+    add_subdirectory(server-core)
     if (LLAMA_BUILD_SERVER)
         add_subdirectory(server)
     endif()
