+++ tools/server-core/README.md
@@ -1,0 +1,3 @@
+# LLaMA.cpp Server core
+
+Core of abstract server for **llama.cpp**.
