+++ tools/server-core/server-core.cpp
@@ -1,0 +1,3 @@
+#include "server-core.h"
+
+// placeholder, some implementation can be moved here
