+++ tools/server-core/server-core.h
@@ -1,0 +1,3 @@
+#pragma once
+
+// placeholder, external API could be moved here
